enterprise/search move to /search
