+++ authentik/core/api/groups.py
@@ -267,7 +267,7 @@ class GroupViewSet(UsedByMixin, ModelViewSet):
     def get_ql_fields(self):
         from djangoql.schema import BoolField, StrField
 
-        from authentik.enterprise.search.fields import (
+        from authentik.search.fields import (
             JSONSearchField,
         )
