+++ authentik/root/settings.py
@@ -105,6 +105,7 @@ TENANT_APPS = [
     "authentik.providers.scim",
     "authentik.rbac",
     "authentik.recovery",
+    "authentik.search",
     "authentik.sources.kerberos",
     "authentik.sources.ldap",
     "authentik.sources.oauth",
@@ -208,17 +209,18 @@ SPECTACULAR_SETTINGS = {
         "authentik.api.v3.schema.response.postprocess_schema_responses",
         "authentik.api.v3.schema.query.postprocess_schema_query_params",
         "authentik.api.v3.schema.cleanup.postprocess_schema_remove_unused",
+        "authentik.search.schema.postprocess_schema_search_autocomplete",
         "authentik.api.v3.schema.enum.postprocess_schema_enums",
     ],
 }
 
 REST_FRAMEWORK = {
-    "DEFAULT_PAGINATION_CLASS": "authentik.api.pagination.Pagination",
+    "DEFAULT_PAGINATION_CLASS": "authentik.search.pagination.AutocompletePagination",
     "DEFAULT_FILTER_BACKENDS": [
+        "authentik.search.ql.QLSearch",
         "authentik.rbac.filters.ObjectFilter",
         "django_filters.rest_framework.DjangoFilterBackend",
         "rest_framework.filters.OrderingFilter",
-        "rest_framework.filters.SearchFilter",
     ],
     "DEFAULT_PERMISSION_CLASSES": ("authentik.rbac.permissions.ObjectPermissions",),
     "DEFAULT_AUTHENTICATION_CLASSES": (
